handle clicks in child views, with parent views notified of state changes via callbacks.
When displaying a shout (the shout, the action bar, the comments,
etc.) in a listview, handle clicks on the various items at the lowest
child view.  Some parent views need to change in response to changes
in the child (e.g., expand the top and bottom margins when the action
bar or comments are displayed). To handle this, state change listener
callbacks can be registered with the child views (e.g.,
ActionShoutView) to be called when the child view state changes.

This commit adopts this callback scheme for the ActionShoutView and
ShoutListViewRow.  The ShoutListViewRow registers with the
ActionShoutView to be notified when the action bar visibility changes.
The row view can then expand the top and bottom margins when the
action bar is visibile and shrink them when it is not.