IdManager manages the user for this app (Me)
SignatureUtility is entirely static and updated to take UnsignedShouts.
Relevant changes were made in other parts of the code to get everything
to compile. The SimpleUnsignedShout is a bit of a hack, and could
probably be cleaned up. Whoever redoes the packet should follow through
the entire ShoutCreation and send chain.