Shout serialization passes initial tests
The issue was that I was copying the wrong range out of the total buffer
over the entire packet body for the parents, which would therefore lead
to the wrong data being put into the hash function. Keeping track of
where a Shout ends and a parent starts fixes this.