pass sender identity into shout creation methods.
Previously, the ShoutCreator constructor created an IdManager
instance, which was used to retrieve the identity for sending.  This
commit removes that IdManager reference and takes the actual
dependency, the signing identity Me, as a parameter of each method.

This make error handling easier (ShoutCreator can assume the passed
identity exists), testing easier (constructing IdManager before make
unit testing difficult), and will make support for multiple sending
identities easier to add in the future.