decouple the create shout tasks from the send task.
Previously, the ShoutTask and CommentTask classes called SendShoutTask
when the shout was created.  This made displaying intermediate errors
(e.g., a failure to create the Shout) to the user difficult.

Now, the MessageActivity class is responsible for iniating the two
tasks separately. If the creation fails, the user is informed and the
activity is kept open so that the message is not lost.  If the
creation succeeds, then SendShoutTask is executed.