pass Me user into shout tasks, rather than pulling from IdManager.
Retrieving the Me user from IdManager in an async task is problematic
because the user might not have setup their identity yet.  The error
handling is complicated.

This commit adds the Me parameter to the async task constructor
instead. The async tasks are usually created by Activities, so they
can retrieve the Me value safely, as they can communicate to the user
that a username is necessary to perform the requested action.