pass NetworkInterface into tasks, rather than constructing internally.
NetworkInterface#unbind() should be called before the NetworkInterface
is destroyed, but currently this is not done.  This commit changes the
usage of NetworkInterface instances in preparation for proper
lifecycle management.

Creating a NetworkInterface takes time (as it must bind to the
NetworkService in the background).  Thus, creating and destroying in
SendShoutTask is non-idle.  Instead, the NetworkInterface is created
in the Activity that will later start SendShoutTask, so that the
service is likely already bound when the task starts.

Proper life-cycle management of the NetworkInterface can be enabled by
linking it to the Activity lifecycle callback methods.