call accepted() method of AgreementListener if user had previously accepted.
This commit modified the getConsent() method to call the
listener.accepted() method if the user had already accepted and thus
the dialog box is not shown.  With this change, one of the two
callback methods, declined() or accepted(), is always called and thus
the calling logic can use them for setup.  The application logic does
not need to be aware of whether the dialog box was shown.