switch NetworkService from Messenger to AIDL-defined binder.
The previous messenger-based interface was clunky and not-type-safe
when the service exposes more than one operation.  In preparation for
implementing an operation that returns whether the service is
initialized, this commit switches to the cleaner AIDL-based interface.

Currently, the service is only accessed from its own process, so IPC
is not needed and a local service (non-AIDL binder) could be used.
But, we might want to allow third-party front-ends later, so we stick
with IPC for service access.