use error codes only at the send() IPC boundary and exceptions elsewhere.
The ErrorCode enum is useful for passing the cause of failure (or
success) of a NetworkInterface#send() call across the service IPC
boundary.  The rest of the Java code (on either side), though, should
be using exceptions.  This commit modifies the callers of send() to
use exceptions, passed between AsyncTasks in type-safe manner thanks
to SendResult.

The visiblity of ErrorCode is changes to package-private to ensure it
is not used elsewhere.