always finish() if the Manes installation or registration was canceled.
Previously, the initiator of the installation or registration activity
should decide if the activity should exit if the request was canceled.
However, this feature requires a non-trivial amount of code (and the
associated complexity).

Always finishing the activity is ok in every situation except when the
user has already composed a shout (then the shout would be lost).  But
getting to that state requires by-passing the regsitration request
when MessageActivity starts, which shouldn't be possible.

Thus, the additional code complexity is not worth the
(never-to-be-used) benefits. This commit removes the option.
Canceling registration or installation will always finish() the
requesting Activity.