add avatar and location fields to Shout serialization format.
This commit adds avatar and longitude-latitude headers to the shout
network format.  Those two features are implemented, in this version
the avatar hash is set to zero and the header flag indicating if
locations are included is set to false.  Incoming packets are parsed
properly, but the avatar hash and locations, if included, are ignored.

This commit also reorganizes the serialization and deserialization
code.  The primary change is to prefer ByteBuffers over arrays, to
avoid lots of memory copies.