switch to Bouncycastle lightweight API
This commit switches all the crypto to use the Bouncycastle
lightweight API, instead of the Java JCA/JCE provider stuff for three
primary reasons:

1) The Java APIs made it nearly impossible to encode signatures in a
   form other than the default ASN.1.  With the lightweight API, it is
   easy to encode the R and S compontents of DSA signatures as we see
   fit.

2) At some point we may want to switch to a C crypto library using
   NDK. The new code uses the lightweight API behind a simple facade,
   so it will be easy to switch the backend if we go the NDK route
   later.

3) The Java API is a complicated mess that requires lots of unsafe
   casting and catching lots of 'impossible' exceptions.  The
   lightweight API is much less complex, at the <sarcasim> "cost" of
   having to modify more code if we every switch from public-key
   signing to authentication based on shared secrets </sarcasm>.

With this commit, all the crypto stuff is in new package
org.whispercomm.shout.crypto.

This commit also modifies the encoding of the signatures in the shout
serialization format.  Instead of ASN.1, they are encoded as two fixed
length unsigned integers, R and S.