implement multi-object Shout packet format
This commit changes the main Shout packet format to be
content-agnostic.  Each packet can contain an arbitrary number of
objects, identified by an object type code.

Shouts are encoded as objects of type 0 and a chain of up to 3 shouts
(child, parent, grandparent) are sent by placing up to 3 objects in
the Shout packet (the child, the parent, and the grandparent objects).