fix bad colors in Preferences title and summary fields
The colors used for the background and text in the Preferences in the
Settings activity are too similar and hard to read.  The current code
overrides the text color programatically. Adds a theme to the Settings
app that sets the text color to black and gray, to contrast with the
white background.