add delegated preferences framework
The provided preferences framework does not allow for external
activities to be used to select the new value for a preference.  For
example, it is not possible to use the Gallery or Camera apps to
select an image for a preference storing a bitmap.  This commit adds a
framework for delegating value selection from the preference to
another object.