add interface for referencing objects by hash
Many objects in Shout are referenced from other objects by hash. For
example, shouts reference avatars by hash and parent shouts by hash.
In some cases, the referenced object may not be available. For
example, avatars are shared asynchronously and without any guarantee.
Conseuqently, the Java classes representing such objects cannot
directly contain the referenced object.

This commit adds a HashReference interface, which contains methods to
access that hash specifying the reference, determine if the referant
is available, and retrieve it if so.  An object with a hash-based
reference to another should use a HashReference implementation to
provide safe access to both the hash and, when available, the
referant.

A default implementation, SimpleHashReference, is also included.