reenable auto saving of MessageFragment UI state on rotate
This fixes the bug where the message text was cleared on rotate or
other configuratin changes.  The framework can only save the UI state
automatically if each view has a unique id.  This was missing from the
fragment view for message activity. This commit specifies a unique id,
reenabling the auto-save.