Allow changing of username in database
Alter the unique column constaint to be over (Username, Public Key) in
the user table. Add a reference to the primary key of the user table
as a column in the Shout table.

This commit also includes some refactoring of the unit test code, done
while updating the tests for the main changes in this commit.