add avatar hash reference to User interface
This commit adds a HashReference<Avatar> object to the User interface
and updates the content provider and network serialization code to use
it.

Newly created Shouts use a dummy all-zero hash, since the real hashes
are not available yet.  That will be fixed when the merkle hashing
code is added.