add two new constructors to Avatar
This commit adds two constructors to Avatar.

The first takes a Bitmap and generates a PNG encoding for saving and
transmission.

The second takes a byte array and mime type.  If the encoded data is
not of that mime type (even if it is a valid image type), an exception
is thrown.  This constructor is useful when decoding Avatars from a
ContentDescriptor for rejecting one with an incorrect mime type.