add object type to canonical form of shouts
Hashable objects are uniquely identified by their hashes, but
different object types may have instances with the same serialization
and thus same short.  To prevent this and to make it possible to infer
an object type from the serialized form, this commit adds an object
header, consisting of the object type and content length, to the shout
canonical form.  The content length is included to make the style more
uniform for future objects (e.g., content sharing).