add implemenation of simple content sharing protocol
This commit adds an implementation of a simple content sharing
protocol (but does not enable it).

The protocol is very simple:

If a content block is missing, a request is sent.  Any neighbor that
hears the request and hash the block will respond with the block and
any child blocks.  If no response is received, the requestor will try
again using exponential backoff up to 7 times, where the backoff is
computed as 3^(#attempts) seconds.

If a node waiting to response to a request hears a response from
someone else, it will cancel its own response.