retrieve missing avatars when shout is first received
Previously, missing avatar content blocks were not requested until a
shout was viewed. This might be long after the sender has left the
area; the avatar should be retreived when the shout is first seen.

The content request handler listens to incoming shouts and tries to
retrieve each avatar, triggering then retrieval if one is
missing. This implemenation is inefficient and ugly; it will be
replaced when avatar access is moved into a content provider.