do not change key when changing username
Since public key are essentially invisible to users right now, it
doesn't make sense to pester them with warnings about keys when
changing usernames.  Instead, we keep the key unchanged.

Later, when we have added some UI that is key related, when can add an
option to change the key.