do not use reference-counted wakelocks with acquireTimeout()
Android's implementation of PowerManager.WakeLock is buggy when
reference counting and acquireTimeout() are used together. Since we
want to use the acquireTimeout() variant to reduce the impact of bad
code on battery life, this commit changes AlarmExecutorService to use
a different wakelock for each task, instead of a shared
reference-counted wakelock.