reduce overdraw in main listview of shouts
The main list of shouts had severe overdraw due to 2 layers of
invisible background, the listview background, and the shoutviews
themselves.  This commit reduces the overdraw by removing the two
hidden layers of background and moving the definition of the gray
background from the listview to the window itself.  Now the overdraw
is only 1X, the white shoutview on top of the gray background.