Fixed top preference category color
The top preference category was a custom widget in order to round the corners
(as far as I know, there's no other way to set the background of the category
as a drawable, but if there is, go ahead and do it). The drawable that was
set as the background did not explicitly set the background to black.