Temporarily removed reshout button from ActionBar
Implementing reshouting from the ActionBar required a change in a method interface. This caused
a bug when clicking the reshout button from the main shout page, since that did not use
the new interface. The reshout button has been temporarily removed from the ActionBar until
a different way of refreshing the Shout is implemented