use shorter form for absolute dates in shout view
Previously, the full "MMM d, yyyy at HH:mm a" form was used for dates
older than two weeks in the shout view.  This is too lengthy on narrow
phones, so this commit removes the time for those old shouts.  The
full date and time is still available in the details view.