modify back button behavior on details to always return to timeline
Previously, when the details activity was launched via a notification,
clicking the back button would return to the task running when the
notification was pressed.  This behavior is not compliant with Android
specification from 3.0 onward.  Instead, the back button should
navigate back in the Shout task, which for us means returning the
timeline view.