separate provider uris from database schema
The content provider URIs do not need to map one-to-one to the
implementing sql schema.  In our case, shout URIs should return a
denormalized view including the author information, so that a separate
query is not needed to retrieve it (in rapidly scrolling lists of
shouts, the extra queries visibily hurt performance).