fix incorrect styling on avatar preference
Custom preferences did not inherit the platform-default styling. This
manifested as incorrect text size and margins in the avatar
preference.

The custom preference styles cannot inherit from the platform
defaults, which are not exposed.  Thus, the custom styles are not
passed to the parent Preference class, which then must use the
platform default style.  The attributes which should be overridden in
the custom style are manually extracted and applied in the
DelegatedPreference constructor.