fix lint errors stemming from api version-specific styling attributes
Under the most-recent Android Lint, specifying an attribute from a
newer API version in an XML file that will be applied on an older API
version is an error.  This means that the ActionBarSherlock trick of
specifying both the Android-namespace attribute from the new API and
the ABS-namespace backwards-compatible attribute, e.g.,
android:actionBarStyle and actionBarStyle, no longer works.

Instead, the attributes must be declared in version-specific styles.
values-v11/styles.xml uses the Android-namespace attributes,
overriding the default ABS-namespace attributes in values/styles.xml.

This commit applies this separation.