add viewport layout
The ViewportLayout can take its size (width, height, or both) from the
runtime size of another view in the hierarchy, not necessarily its own
parent.

For example, this is useful for matching the height of an view inside
of a ScrollView to the visible portion (or viewport) of the
ScrollView.