use stored context in shout list loader
The anonymous CursorLoader implementation used to call the parent
ShoutListFragment's getActivity() method in its onCreateLoader()
callback to retrieve a context.  Sometimes when navigating back to the
fragment, getActivity() returns null, resulting in an NPE in the
loader.

The CursorLoader already has a Context supplied through its
constructor, so use that in onCreateLoader() instead.