deletion button contacting content provider
Add classes and methods to help establish connection between the
deletion button and content provider. Changes are saved and stored
in the content provider

ShoutEraser class is added

A helper class to delete shout and contact the content provider.

DeleteTask class is added

Task of deleting shout is created on touch of deletion button. Changes
will be stored in content provider.

Deletion connecting button to content provider