do not load avatar from disk in LocalUserImpl
LocalUserImpl is often constructed and accessed on the main thread, so
it should not touch the disk.  With this commit, the reference
returned by #getAvatar() will always return a simple hash reference
without the content loaded.  Users can then query the ImageProvider in
main-thread-safe fashion to get the actual content, if available.