delay decoding public key until actually needed
The LocalUserImpl object decodes the public key on construction.  This
operation is relatively expensive and causes garbage collection churn,
particulary with LocalUserImpl objects are being created during fast
scrolling of a listview.

Usually the public key field is never accessed, so this commit changes
LocalUserImpl to delay decoding the public key until is actually used.